Task(Yandex)/app/requirements.txt
django==5.0
djangorestframework==3.15.2
psycopg2-binary~=2.0
django-environ==0.11.2
drf-spectacular==0.27.2
requests==2.32.3